Google Gemini Assignment/requirements.txt
google-generativeai>=0.3.2
beautifulsoup4>=4.12.2
trafilatura>=1.6.3
lxml>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0

